+++ lib/db.bash
@@ -69,24 +69,22 @@ function odoo_db_exists {
     fi
 }
 
-# odoo_db_dump <dbname> [format|odoo_conf_file]
-# if second argument is file and it exists, then it used as config filename
-# in other cases second argument is treated as format
+# odoo_db_dump <dbname> <file-path> [format [odoo_conf_file]]
+# dump database to specified path
 function odoo_db_dump {
-    local FILE_SUFFIX=`date -I`.`random_string 4`;
     local db_name=$1;
-    local db_dump_file="$BACKUP_DIR/db-backup-$db_name-$FILE_SUFFIX.backup";
+    local db_dump_file=$2;
     local conf_file=$ODOO_CONF_FILE;
 
-    if [ -f "$2" ]; then
-        conf_file=$2;
-    else
-        local format=$2;
+    if [ -f "$3" ]; then
+        conf_file=$3;
+    elif [ ! -z $3 ]; then
+        local format=$3;
         local format_opt=", '$format'";
-    fi
 
-    if [ ! -z $format ]; then
-        db_dump_file="$db_dump_file.$format";
+        if [ -f "$4" ]; then
+            conf_file=$4;
+        fi
     fi
 
     local python_cmd="import erppeek; cl=erppeek.Client(['-c', '$conf_file']);";
@@ -97,9 +95,26 @@ function odoo_db_dump {
         echov "Database named '$db_name' dumped to '$db_dump_file'!";
         return 0;
     else
-        echov "Database '$db_name' fails on dump!";
+        echo "Database '$db_name' fails on dump!";
         return 1;
     fi
+}
+
+
+# odoo_db_backup <dbname> [format [odoo_conf_file]]
+# if second argument is file and it exists, then it used as config filename
+# in other cases second argument is treated as format, and third (if passed) is treated as conf file
+function odoo_db_backup {
+    if [ -z $BACKUP_DIR ]; then
+        echo "Backup dir is not configured. Add 'BACKUP_DIR' variable to your 'odoo-helper.conf'!";
+        return 1;
+    fi
+
+    local FILE_SUFFIX=`date -I`.`random_string 4`;
+    local db_name=$1;
+    local db_dump_file="$BACKUP_DIR/db-backup-$db_name-$FILE_SUFFIX.backup";
+
+    odoo_db_dump $db_name $db_dump_file $2 $3;
 }
 
 # odoo_db_restore <dbname> <dump_file> [odoo_conf_file]
@@ -129,7 +144,8 @@ function odoo_db_command {
         $SCRIPT_NAME db exists <name> [odoo_conf_file]
         $SCRIPT_NAME db create <name> [odoo_conf_file]
         $SCRIPT_NAME db drop <name> [odoo_conf_file]
-        $SCRIPT_NAME db dump <name> [format|odoo_conf_file]
+        $SCRIPT_NAME db dump <name> <dump_file_path> [format [odoo_conf_file]]
+        $SCRIPT_NAME db backup <name> [format [odoo_conf_file]]
         $SCRIPT_NAME db restore <name> <dump_file_path> [odoo_conf_file]
 
     ";
@@ -163,6 +179,11 @@ function odoo_db_command {
                 odoo_db_dump "$@";
                 exit;
             ;;
+            backup)
+                shift;
+                odoo_db_backup "$@";
+                exit;
+            ;;
             restore)
                 shift;
                 odoo_db_restore "$@";
@@ -170,7 +191,6 @@ function odoo_db_command {
             ;;
             exists)
                 shift;
-                local VERBOSE=1;
                 odoo_db_exists "$@";
                 exit;
             ;;